+++ front/src/App/MyTalks/TalkTable/components/ActionsColumn/ScheduleButton.js
@@ -1,0 +1,14 @@
+import ButtonLoading from '#shared/ButtonLoading';
+import { ScheduleIcon } from '#shared/icons';
+import React from 'react';
+
+export function ScheduleButton(props) {
+  return (
+    <ButtonLoading
+      icon={<ScheduleIcon />}
+      color={'transparent'}
+      onClick={props.onClick}
+      tooltipText={'Agendar'}
+    />
+  );
+}
